% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_halogen_app}
\alias{run_halogen_app}
\title{Run the HaloGen Shiny Application}
\usage{
run_halogen_app()
}
\value{
Does not return a value. This function is called for its side effect of launching the Shiny application.
}
\description{
This function launches the main user interface for the HaloGen application,
allowing for Likelihood Ratio calculation based on DNA quantities.
}
